renamed messages.js to Message.js
