+++ src/util.js
@@ -25,9 +25,9 @@ const getEmoji = (search) =>
 
 /**
  * @param {string} name
- * @returns {Promise<import('discord.js').Channel>}
+ * @returns {import('discord.js').Channel}
  */
-async function getChannel(name) {
+function getChannel(name) {
     return client.channels.cache.get(config[name]);
 }
